requirements.txt
Django==3.1
psycopg2-binary==2.8
djangorestframework
markdown
django-filter
